Run containers as the uid and gid of the user
There are two main reasons why we don't want to run as root inside the
containers:

1. Running as root is bad practice and a security risk.
2. All files generated within the container (eg `__pycache__` and
   artifacts etc) are owned by the root user on the host, resulting in
   file ownership issues in the user's project directory.

My approach in this commit is to create Docker images that execute
commands as the user running `hatch`.

In the vast majority of cases the user will be the only user on the
system, but on shared machines we need to handle this gracefully. My
solution is to treat every UID:GID combination as a unique identity and
create a Docker image with a name that looks like this:

  python_3.11-alpine_1000_1000

In this case the UID and GID are both `1000`. Every time this particular
user executes `hatch run` for this environment, `hatch-containers` will
select this container (if it already exists).

This fixes:

- https://github.com/ofek/hatch-containers/issues/12
- https://github.com/ofek/hatch-containers/issues/13